requirements.txt
flask
transformers
torch
nltk
gunicorn
python-dotenv
requests
scikit-learn